atu: Move UAPI header into "misc/"
UAPI headers need to be in a directory to be installed using
"make headers_install" for userspace access.
