+++ include/uapi/misc/Kbuild
@@ -1,2 +1,3 @@
 # misc Header export list
+header-y += atu_ioctl.h
 header-y += cxl.h
